pyproject.toml
[tool.poetry]
name = "kriging-interpolation"
version = "0.2.0"
description = "Api for working with interpolation kriging methods"
authors = ["Askar akhmadullin <akhmadullin.a@internet.ru>"]
readme = "README.md"
packages = [{include = "kriging_interpolation"}]

[tool.poetry.dependencies]
python = "3.12.1"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"
pydantic-settings = "^2.1.0"
pykrige = "^1.7.1"
matplotlib = "^3.8.2"
pydantic-extra-types = "^2.2.0"
scikit-learn = "^1.3.2"
gstools = "^1.5.1"
cartopy = "^0.22.0"
notebook = "^7.0.6"
motor = "^3.3.2"
celery = {extras = ["redis"], version = "^5.3.6"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
